+++ singlefactorauth/build.gradle
@@ -38,7 +38,7 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
 
-    implementation 'org.web3j:crypto:4.9.5'
+    implementation 'org.web3j:crypto:5.0.0'
     implementation 'org.torusresearch:torus-utils-java:4.0.0'
     implementation 'org.torusresearch:fetch-node-details-java:5.0.0'
     implementation 'com.google.code.gson:gson:2.10.1'
