+++ singlefactorauth/src/main/java/com/web3auth/singlefactorauth/types/ErrorCode.kt
@@ -1,0 +1,9 @@
+package com.web3auth.singlefactorauth.types
+
+enum class ErrorCode {
+    USER_ALREADY_ENABLED_MFA,
+    PRIVATE_KEY_NOT_FOUND,
+    RUNTIME_ERROR,
+    SOMETHING_WENT_WRONG,
+    CONTEXT_NOT_FOUND
+}
