+++ singlefactorauth/src/main/java/com/web3auth/singlefactorauth/types/SFAError.kt
@@ -21,11 +21,3 @@ object SFAError {
         }
     }
 }
-
-enum class ErrorCode {
-    USER_ALREADY_ENABLED_MFA,
-    PRIVATE_KEY_NOT_FOUND,
-    RUNTIME_ERROR,
-    SOMETHING_WENT_WRONG,
-    CONTEXT_NOT_FOUND
-}
