+++ singlefactorauth/src/main/java/com/web3auth/singlefactorauth/types/SingleFactorAuthArgs.kt
@@ -5,15 +5,17 @@ import org.torusresearch.fetchnodedetails.types.Web3AuthNetwork
 class SingleFactorAuthArgs(network: Web3AuthNetwork, clientid: String, networkUrl: String? = null, serverTimeOffset: Int = 0) {
     private var network: Web3AuthNetwork
     private var clientid: String = "torus-default"
-    private var networkUrl: String? = null
+    private var networkUrl: String? = networkUrl
     private var enableOneKey: Boolean
-    private var serverTimeOffset: Int = 0
+    private var serverTimeOffset: Int = serverTimeOffset
 
 
     init {
         this.network = network
         this.clientid = clientid
         this.enableOneKey = true
+        this.networkUrl = networkUrl
+        this.serverTimeOffset = serverTimeOffset
     }
     fun getClientId(): String? {
         return clientid
